refactor: service command
