+++ cli/app/commands/service/messages.py
@@ -1,32 +1,3 @@
-dry_run_mode = "=== DRY RUN MODE ==="
-dry_run_command_would_be_executed = "The following commands would be executed:"
-dry_run_command = "Command:"
-dry_run_service = "Service:"
-dry_run_detach_mode = "Detach mode:"
-dry_run_env_file = "Environment file:"
-end_dry_run = "=== END DRY RUN ==="
-starting_services = "Starting services: {services}"
-services_started_successfully = "Services started successfully: {services}"
-service_start_failed = "Service start failed: {error}"
-unexpected_error_during_start = "Unexpected error during start: {error}"
-stopping_services = "Stopping services: {services}"
-services_stopped_successfully = "Services stopped successfully: {services}"
-service_stop_failed = "Service stop failed: {error}"
-unexpected_error_during_stop = "Unexpected error during stop: {error}"
-checking_services = "Checking status of services: {services}"
-services_status_retrieved = "Services status retrieved successfully: {services}"
-service_status_failed = "Service status check failed: {error}"
-unexpected_error_during_status = "Unexpected error during status check: {error}"
-restarting_services = "Restarting services: {services}"
-services_restarted_successfully = "Services restarted successfully: {services}"
-service_restart_failed = "Service restart failed: {error}"
-unexpected_error_during_restart = "Unexpected error during restart: {error}"
-unknown_error = "Unknown error occurred"
-service_action_info = "{action} services: {name}"
-service_action_success = "Service {action} successful: {name}"
-service_action_failed = "Service {action} failed: {error}"
-service_action_unexpected_error = "Unexpected error during {action}: {error}"
-environment_file_not_found = "Environment file not found: {path}"
 compose_file_not_found = "Compose file not found: {path}"
 docker_command_executing = "Executing Docker command: {command}"
 docker_command_completed = "Docker command completed successfully for {action} action"
@@ -34,5 +5,7 @@ docker_command_failed = "Docker command failed with return code {return_code}"
 docker_command_stdout = "Docker command stdout: {output}"
 docker_command_stderr = "Docker command stderr: {output}"
 docker_unexpected_error = "Unexpected error during {action} action: {error}"
-command_output_label = "Command output: {output}"
-command_error_label = "Command error: {output}"
+environment_file_not_found = "Environment file not found: {path}"
+service_action_info = "{action} services: {name}"
+service_action_failed = "Service {action} failed: {error}"
+service_action_unexpected_error = "Unexpected error during {action}: {error}"
