+++ cli/app/main.py
@@ -8,7 +8,6 @@ from rich.text import Text
 
 from app.commands.clone.command import clone_app
 from app.commands.install.command import install_app
-from app.commands.service.command import service_app
 from app.commands.test.command import test_app
 from app.commands.uninstall.command import uninstall_app
 from app.commands.update.command import update_app
@@ -87,7 +86,6 @@ def main(
 
 
 app.add_typer(clone_app, name="clone")
-app.add_typer(service_app, name="service")
 app.add_typer(install_app, name="install")
 app.add_typer(uninstall_app, name="uninstall")
 app.add_typer(update_app, name="update")
